regint/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary]
python-multipart
python-dotenv
pydantic<3
